Reworked how realm's initramfs is provided
It's now shipped as tar that is unpacked on init_out.
During build (new step, build_initramfs) it's packaged and put in
place. This way it's easier to modify the boot process or add some
persistent tools to it.

Its provided content also changed a little. 9P is mounted
automatically under /qemu. So the same shared dir is visible for main
host, the FVP host and the realm.