Removed optee-build, manual rootfs/initramfs preparation
Optee-build has been removed. Everything it did is now done inside fvp-cca.sh
script. Kernel is built by hand. Initramfs-host is now done in the same way as
initramfs-realm. It's unpacked in the init stage and packed in build stage so
it's easy to temporarily modify its content if needed.